Resolves issue with INT_MIN limit by creating a long int variable \num_long

+++ main.h
@@ -23,7 +23,7 @@ int specify_sub(char spec, va_list values);
 int _putchar(char c);
 int _malloc(va_list num_list, int base);
 int _malloc2(va_list num_list, int base);
-int fill_buffer(char *buffer, int num, int n, int base);
+int fill_buffer(char *buffer, long int num, int n, int base);
 int print_rev_buffer(char *buffer, int n);
 
 int _malloc3(va_list num_list, int base, char caps);
